+++ besu-plugins/linea-sequencer/sequencer/src/main/java/net/consensys/linea/sequencer/txpoolvalidation/LineaTransactionPoolValidatorFactory.java
@@ -80,11 +80,10 @@ public class LineaTransactionPoolValidatorFactory implements PluginTransactionPo
   public PluginTransactionPoolValidator createTransactionValidator() {
     final var validators =
         new PluginTransactionPoolValidator[] {
-          new AllowedAddressValidator(denied, rejectedTxJsonRpcManager),
-          new GasLimitValidator(txPoolValidatorConf, rejectedTxJsonRpcManager),
-          new CalldataValidator(txPoolValidatorConf, rejectedTxJsonRpcManager),
-          new ProfitabilityValidator(
-              besuConfiguration, blockchainService, profitabilityConf, rejectedTxJsonRpcManager),
+          new AllowedAddressValidator(denied),
+          new GasLimitValidator(txPoolValidatorConf),
+          new CalldataValidator(txPoolValidatorConf),
+          new ProfitabilityValidator(besuConfiguration, blockchainService, profitabilityConf),
           new SimulationValidator(
               blockchainService,
               transactionSimulationService,