+++ besu-plugins/linea-sequencer/sequencer/src/main/java/net/consensys/linea/sequencer/txselection/selectors/LineaTransactionSelector.java
@@ -14,10 +14,15 @@
  */
 package net.consensys.linea.sequencer.txselection.selectors;
 
+import static net.consensys.linea.sequencer.txselection.LineaTransactionSelectionResult.TX_MODULE_LINE_COUNT_OVERFLOW;
+import static net.consensys.linea.sequencer.txselection.LineaTransactionSelectionResult.TX_MODULE_LINE_COUNT_OVERFLOW_CACHED;
+
 import java.time.Instant;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 
 import lombok.extern.slf4j.Slf4j;
 import net.consensys.linea.config.LineaProfitabilityConfiguration;
@@ -41,6 +46,7 @@ public class LineaTransactionSelector implements PluginTransactionSelector {
   private TraceLineLimitTransactionSelector traceLineLimitTransactionSelector;
   private final List<PluginTransactionSelector> selectors;
   private final Optional<JsonRpcManager> rejectedTxJsonRpcManager;
+  private final Set<String> rejectedTransactionReasonsMap = new HashSet<>();
 
   public LineaTransactionSelector(
       final BlockchainService blockchainService,
@@ -51,6 +57,13 @@ public class LineaTransactionSelector implements PluginTransactionSelector {
       final Map<String, Integer> limitsMap,
       final Optional<JsonRpcManager> rejectedTxJsonRpcManager) {
     this.rejectedTxJsonRpcManager = rejectedTxJsonRpcManager;
+
+    // only report rejected transaction selection result from TraceLineLimitTransactionSelector
+    if (rejectedTxJsonRpcManager.isPresent()) {
+      rejectedTransactionReasonsMap.add(TX_MODULE_LINE_COUNT_OVERFLOW.toString());
+      rejectedTransactionReasonsMap.add(TX_MODULE_LINE_COUNT_OVERFLOW_CACHED.toString());
+    }
+
     selectors =
         createTransactionSelectors(
             blockchainService,
@@ -163,7 +176,8 @@ public class LineaTransactionSelector implements PluginTransactionSelector {
 
     rejectedTxJsonRpcManager.ifPresent(
         jsonRpcManager -> {
-          if (transactionSelectionResult.discard()) {
+          if (transactionSelectionResult.discard()
+              && rejectedTransactionReasonsMap.contains(transactionSelectionResult.toString())) {
             jsonRpcManager.submitNewJsonRpcCallAsync(
                 JsonRpcRequestBuilder.generateSaveRejectedTxJsonRpc(
                     jsonRpcManager.getNodeType(),