+++ besu-plugins/linea-sequencer/sequencer/src/test/java/net/consensys/linea/sequencer/txpoolvalidation/validators/CalldataValidatorTest.java
@@ -38,8 +38,7 @@ public class CalldataValidatorTest {
         new CalldataValidator(
             LineaTransactionPoolValidatorCliOptions.create().toDomainObject().toBuilder()
                 .maxTxCalldataSize(MAX_TX_CALLDATA_SIZE)
-                .build(),
-            Optional.empty());
+                .build());
   }
 
   @Test