+++ besu-plugins/linea-sequencer/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Next release
+* feat: Report rejected transactions only due to trace limit overflows to an external service.
 * feat: Report rejected transactions to an external service for validators used by LineaTransactionPoolValidatorPlugin [#85](https://github.com/Consensys/linea-sequencer/pull/85)
 * feat: Report rejected transactions to an external service for LineaTransactionSelector used by LineaTransactionSelectorPlugin [#69](https://github.com/Consensys/linea-sequencer/pull/69)